+++ R/typical.R
@@ -1,12 +1,13 @@
 #' Find the typical value
 #'
-#' For numeric vectors, it returns the median. For factors, characters, and
-#' logical vectors, it returns the most frequent value. If multiple values are
-#' tied for most frequent, it returns them all. \code{NA} missing values are
-#' always silently dropped.
+#' For numeric, integer, and ordered factor vectors, it returns the median.
+#' For factors, characters, and logical vectors, it returns the most
+#' frequent value. If multiple values are tied for most frequent, it returns
+#' them all. \code{NA} missing values are always silently dropped.
 #'
 #' @param x A vector
 #' @export
+#' @importFrom stats quantile
 #' @examples
 #' # median of numeric vector
 #' typical(rpois(100, lambda = 10))
@@ -19,6 +20,10 @@
 #' # if tied, returns them all
 #' x <- c("a", "a", "b", "b", "c")
 #' typical(x)
+#'
+#' # median of an ordered factor
+#' typical(ordered(c("a", "a", "b", "c", "d")))
+#'
 typical <- function(x) {
   UseMethod("typical")
 }
@@ -43,4 +48,14 @@ typical.character <- function(x) {
 #' @export
 typical.logical <- function(x) {
   mean(x, na.rm = TRUE) >= 0.5
+}
+
+#' @export
+typical.integer <- function(x) {
+  unname(stats::quantile(x, 0.5, type = 1, na.rm = TRUE))
+}
+
+#' @export
+typical.ordered <- function(x) {
+  as.character(stats::quantile(x, 0.5, type = 1, na.rm = TRUE))
 }