+++ NEWS.md
@@ -1,5 +1,7 @@
 # modelr 0.1.1.9000
 
+* Added `typical.ordered` and `typical.integer` methods (@jrnold, #44)
+
 # modelr 0.1.1
 
 * Added a `NEWS.md` file to track changes to the package.